Implementing repository service pattern on Login
